working on migrating home page to app directory
