+++ components/layout/footer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
