+++ package.json
@@ -16,12 +16,12 @@
         "bootstrap": "^4.5.3",
         "contentful": "^10.6.15",
         "moment": "^2.29.1",
-        "next": "13",
+        "next": "13.4",
         "next-redux-wrapper": "^6.0.2",
         "react": "^18.2.0",
-        "react-bootstrap": "^1.4.0",
+        "react-bootstrap": "^2.10.0",
         "react-dom": "^18.2.0",
-        "react-markdown": "^5.0.3",
+        "react-markdown": "^9.0.1",
         "react-redux": "^7.2.2",
         "redux": "^4.0.5",
         "redux-devtools-extension": "^2.13.8",
